requirements.txt
#flask
#requests
RPi.GPIO
spidev
azure-iot-device